Added multiprocessing and contextfetching via url
I've added multiprocessing which can be switched on via -mp to speed up the process via pythons multiprocessing tool. Therefore the records won't printed out via stdout, they will be printed out to chunkfiles.
Without -mp, the program-behaviour is the same as before.
The bottleneck of the json.load(file) is still there...
also, if a context_url is set, you dont need to put a context-file over the arguments. the program is going to fetch the context by itself.


Tested with DNB-Jobs and DNB-Adresses.